refactor: extract EffectCards from EffectChain (658→361 lines)
New file: src/components/mixer/EffectCards.tsx (329 lines)
- HSlider, EQ3Card + EQCurve, CompressorCard, ReverbCard
- DelayCard, DistortionCard, FilterCard, EFFECT_COLORS

EffectChain.tsx: 658 → 361 lines (45% reduction)
Build: 0 errors. Tests: 46/46 pass.
